Calculator Commit
🧮 Calculator Project
This Calculator Project provides essential arithmetic operations, including addition, subtraction, multiplication, and division, making it an ideal tool for beginners with a basic understanding of HTML, CSS, and JavaScript. The project features an intuitive user interface, complete with interactive buttons for entering data and a display screen for showing results. Designed with simplicity and functionality in mind, this calculator is a great hands-on project for those looking to enhance their JavaScript skills through practical application.

Key Features:

➕ Addition, ➖ Subtraction, ✖️ Multiplication, ➗ Division
👆 User-friendly buttons for easy input
📟 Display screen for real-time calculation results